requirements.txt
# Core Dependencies
markitdown>=0.0.1a3
pypdf>=3.0.0
python-dotenv>=0.19.0
requests>=2.26.0

# PDF Processing Tools
pdfminer.six>=20221105
pdf2image>=1.16.0
opencv-python>=4.7.0
camelot-py>=0.10.1
ghostscript>=0.7

# Additional Dependencies
spacy
tqdm
